+++ test/notebook_plot.R
@@ -8,9 +8,9 @@ df <- datasets::mtcars
 S <- cor(df)
 
 library(msPCA)
-mspca_results <- msPCA::mspca(S, 2, c(6,6), verbose=TRUE)
+mspca_results <- msPCA::mspca(S, 2, c(6,6), type="Sigma", verbose=TRUE)
 
-mspca_results <- msPCA::mspca(S, 3, c(6,6,6), verbose=TRUE)
+mspca_results <- msPCA::mspca(S, 3, c(6,6,6), type="Sigma", verbose=TRUE)
 
 library(dplyr)
 library(ggplot2)