+++ test/notebook_synthetic.R
@@ -62,12 +62,23 @@ for (iter in 1:5){ #Number of replications
     S <- cov(X)
 
     #msPCA algorithm
-    mspca_results <- msPCA::mspca(S, 2, c(k,k), verbose=FALSE)
+    mspca_results <- msPCA::mspca(S, 2, c(k,k), verbose=FALSE, maxRestartTPM=30)
     resdf[icol,] <- list(n,"msPCA",
                    sum(abs(mspca_results$x_best) > 0), #Sparsity level
                    accuracy(xtrue, mspca_results$x_best), #Support recovery accuracy
                    feasibility_violation_off(S, mspca_results$x_best, 0), #Orthogonality violation
                    fraction_variance_explained(Strue,mspca_results$x_best)) #Fraction of variance explained
+
+    icol <- icol +1
+
+    #msPCA algorithm working on the X matrix directly
+    mspca_X_results <- msPCA::mspca(X, 2, c(k,k), verbose=FALSE, type="X", maxRestartTPM=30)
+
+    resdf[icol,] <- list(n,"msPCA - X",
+                         sum(abs(mspca_X_results$x_best) > 0), #Sparsity level
+                         accuracy(xtrue, mspca_X_results$x_best), #Support recovery accuracy
+                         feasibility_violation_off(S, mspca_X_results$x_best, 0), #Orthogonality violation
+                         fraction_variance_explained(Strue,mspca_X_results$x_best)) #Fraction of variance explained
 
     icol <- icol +1